* fix tests and add store migrations

* add test case

* backport changes

* get tx receipt to check failed tx in nix test

+++ x/vesting/keeper/msg_server.go
@@ -371,6 +371,11 @@ func (k Keeper) ConvertVestingAccount(
 		return nil, errorsmod.Wrapf(errortypes.ErrInvalidRequest, "vesting coins still left in account: %s", msg.VestingAddress)
 	}
 
+	// check if account has any locked up coins left
+	if vestingAcc.HasLockedCoins(ctx.BlockTime()) {
+		return nil, errorsmod.Wrapf(errortypes.ErrInvalidRequest, "locked up coins still left in account: %s", msg.VestingAddress)
+	}
+	
 	// if gov clawback is disabled, remove the entry from the store.
 	// if no entry is found for the address, this will no-op
 	k.DeleteGovClawbackDisabled(ctx, address)
